mysite/requirements.txt
absl-py==1.4.0
asgiref==3.5.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
clang==5.0
Django==3.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.10.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
Markdown==3.3.6
numpy==1.24.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==23.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.7.0
